% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_random_num.R
\name{random_num}
\alias{random_num}
\title{random_num function}
\usage{
random_num()
}
\arguments{
\item{Internal}{parameters for the function.}
}
\value{
The return value, if any, from executing the function.
}
\description{
A shiny function
}
